Split out utilities classes into their own files
Closes delta-io/delta#2351

GitOrigin-RevId: 0864356b9249eea6aa0f290d491d52b05a1e93e1
